# [MRG+1] MAINT: Refactor and speed up silhoutte_score (samples)