A refactor of `metrics.silhouette_score`. Able to get at least 2 times speedup in most cases.

     n_samples=10000, n_labels=6
     In this branch: 13.4s, in master:23.2s
     n_samples=10000, n_labels=4
     In this branch: 14.4s, in master:23.9s

     n_samples=1000, n_labels=6
     In this branch: 115ms, in master:287ms
     n_samples=1000, n_labels=4
     In this branch: 110ms, in master:249ms

     n_samples=100, n_labels=6
     In this branch: 1.87ms, in master:7.13ms
     n_samples=100, n_labels=4
     In this branch: 1.54ms, in master:5.84ms

Also fixed a bug related to non-encoded labels.