@MechCoder 2016-01-10
@GaelVaroquaux I have added comments and replaced a few variable names. Can you tell me if it still harder to follow?