@MechCoder 2016-01-10
In general, I think the tests can be made more stronger but I'll leave that for another PR