@MechCoder 2016-01-13
@TomDLT looks ok now?