@MechCoder 2016-01-13
@agramfort if you have the time :-)