@amueller 2016-01-15
how does it scale to very large ``n_labels``?