@MechCoder 2016-01-15
rebased