@GaelVaroquaux 2016-01-17
:+1: on my side. This is a good rewrite. It is clear and fast. I am merging this!